move all functions to realize system calls to the directory syscalls
